+++ internal/client/network-gossip/validator.go
@@ -1,8 +1,8 @@
 package gossip
 
 import (
-	"github.com/ChainSafe/gossamer/client/network/role"
-	"github.com/ChainSafe/gossamer/primitives/runtime"
+	"github.com/ChainSafe/gossamer/internal/client/network/role"
+	"github.com/ChainSafe/gossamer/internal/primitives/runtime"
 	libp2p "github.com/libp2p/go-libp2p/core"
 )
