+++ internal/client/network/network.go
