+++ internal/primitives/core/ed25519/ed25519.go
